moving routes under commans
