+++ api-client/src/runs/index.ts
@@ -9,7 +9,7 @@ export { getCommand } from './commands/getCommand'
 export { getCommands } from './commands/getCommands'
 export { getCommandsAsPreSerializedList } from './commands/getCommandsAsPreSerializedList'
 export { createRunAction } from './createRunAction'
-export { getRunCommandErrors } from './getRunCommandErrors'
+export { getRunCommandErrors } from './commands/getRunCommandErrors'
 export * from './createLabwareOffset'
 export * from './createLabwareDefinition'
 export * from './constants'
