+++ api-client/src/runs/types.ts
@@ -9,7 +9,6 @@ import type {
   RunTimeParameter,
 } from '@opentrons/shared-data'
 import type { ResourceLink, ErrorDetails } from '../types'
-import type { GetCommandsParams } from './commands/types'
 export * from './commands/types'
 
 export const RUN_STATUS_IDLE = 'idle' as const
@@ -99,11 +98,6 @@ export interface GetRunsParams {
 export interface Runs {
   data: RunData[]
   links: RunsLinks
-}
-
-export interface RunCommandErrors {
-  data: RunCommandError[]
-  meta: GetCommandsParams & { totalLength: number }
 }
 
 export const RUN_ACTION_TYPE_PLAY: 'play' = 'play'
