+++ react-api-client/src/runs/index.ts
@@ -13,7 +13,7 @@ export { useRunActionMutations } from './useRunActionMutations'
 export { useAllCommandsQuery } from './useAllCommandsQuery'
 export { useAllCommandsAsPreSerializedList } from './useAllCommandsAsPreSerializedList'
 export { useCommandQuery } from './useCommandQuery'
-export { useAllRunCommandErrorsQuery } from './useAllRunCommandErrorsQuery'
+export { useRunCommandErrors } from './useRunCommandErrors'
 export * from './useCreateLabwareOffsetMutation'
 export * from './useCreateLabwareDefinitionMutation'
 export * from './useUpdateErrorRecoveryPolicy'
